+++ src/main/java/com/example/ddingsroom/user/dto/ChangeUsernameDTO.java
@@ -6,12 +6,12 @@ import lombok.Setter;
 @Setter
 @Getter
 public class ChangeUsernameDTO {
-    private Integer userId;
+    private Long userId;
     private String newUsername;
 
     public ChangeUsernameDTO() {}
 
-    public ChangeUsernameDTO(Integer userId, String newUsername) {
+    public ChangeUsernameDTO(Long userId, String newUsername) {
         this.userId = userId;
         this.newUsername = newUsername;
     }
